Added onOutDayClick option
onOutDayClick option expects a callback function that will fire off whenever the user clicks on a day cell not currently in the same month in the current view.  The onCalendarRefresh function will also fire off on this event as well so it can still be used whenever the calendar changes the month currently being viewed.